+++ config/gg_handler/gg_data_alas.tmp
@@ -1,0 +1,4 @@
+alas
+gg_on=False
+gg_enable=False
+gg_auto=False
