+++ module/equipment/assets.py
@@ -25,8 +25,16 @@ FLEET_DETAIL = Button(area={'cn': (906, 644, 1008, 699), 'en': (906, 644, 1008, 
 FLEET_DETAIL_CHECK = Button(area={'cn': (915, 647, 980, 679), 'en': (915, 649, 980, 677), 'jp': (915, 647, 979, 679), 'tw': (915, 647, 980, 680)}, color={'cn': (234, 195, 154), 'en': (236, 200, 160), 'jp': (234, 195, 154), 'tw': (237, 196, 154)}, button={'cn': (915, 647, 980, 679), 'en': (915, 649, 980, 677), 'jp': (915, 647, 979, 679), 'tw': (915, 647, 980, 680)}, file={'cn': './assets/cn/equipment/FLEET_DETAIL_CHECK.png', 'en': './assets/en/equipment/FLEET_DETAIL_CHECK.png', 'jp': './assets/jp/equipment/FLEET_DETAIL_CHECK.png', 'tw': './assets/tw/equipment/FLEET_DETAIL_CHECK.png'})
 FLEET_DETAIL_ENTER = Button(area={'cn': (729, 314, 802, 438), 'en': (729, 314, 802, 438), 'jp': (729, 314, 802, 438), 'tw': (729, 314, 802, 438)}, color={'cn': (154, 149, 147), 'en': (154, 149, 147), 'jp': (154, 149, 147), 'tw': (154, 149, 147)}, button={'cn': (729, 314, 802, 438), 'en': (729, 314, 802, 438), 'jp': (729, 314, 802, 438), 'tw': (729, 314, 802, 438)}, file={'cn': './assets/cn/equipment/FLEET_DETAIL_ENTER.png', 'en': './assets/en/equipment/FLEET_DETAIL_ENTER.png', 'jp': './assets/jp/equipment/FLEET_DETAIL_ENTER.png', 'tw': './assets/tw/equipment/FLEET_DETAIL_ENTER.png'})
 FLEET_DETAIL_ENTER_FLAGSHIP = Button(area={'cn': (103, 327, 176, 437), 'en': (103, 327, 176, 437), 'jp': (103, 327, 176, 437), 'tw': (103, 327, 176, 437)}, color={'cn': (141, 135, 138), 'en': (141, 135, 138), 'jp': (141, 135, 138), 'tw': (141, 135, 138)}, button={'cn': (103, 327, 176, 437), 'en': (103, 327, 176, 437), 'jp': (103, 327, 176, 437), 'tw': (103, 327, 176, 437)}, file={'cn': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP.png', 'en': './assets/en/equipment/FLEET_DETAIL_ENTER_FLAGSHIP.png', 'jp': './assets/jp/equipment/FLEET_DETAIL_ENTER_FLAGSHIP.png', 'tw': './assets/tw/equipment/FLEET_DETAIL_ENTER_FLAGSHIP.png'})
+FLEET_DETAIL_ENTER_FLAGSHIP_HARD_1 = Button(area={'cn': (390, 173, 476, 259), 'en': (390, 173, 476, 259), 'jp': (390, 173, 476, 259), 'tw': (390, 173, 476, 259)}, color={'cn': (91, 71, 68), 'en': (91, 71, 68), 'jp': (91, 71, 68), 'tw': (91, 71, 68)}, button={'cn': (390, 173, 476, 259), 'en': (390, 173, 476, 259), 'jp': (390, 173, 476, 259), 'tw': (390, 173, 476, 259)}, file={'cn': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP_HARD_1.png', 'en': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP_HARD_1.png', 'jp': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP_HARD_1.png', 'tw': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP_HARD_1.png'})
+FLEET_DETAIL_ENTER_FLAGSHIP_HARD_2 = Button(area={'cn': (390, 287, 476, 373), 'en': (390, 287, 476, 373), 'jp': (390, 287, 476, 373), 'tw': (390, 287, 476, 373)}, color={'cn': (69, 59, 55), 'en': (69, 59, 55), 'jp': (69, 59, 55), 'tw': (69, 59, 55)}, button={'cn': (390, 287, 476, 373), 'en': (390, 287, 476, 373), 'jp': (390, 287, 476, 373), 'tw': (390, 287, 476, 373)}, file={'cn': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP_HARD_2.png', 'en': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP_HARD_2.png', 'jp': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP_HARD_2.png', 'tw': './assets/cn/equipment/FLEET_DETAIL_ENTER_FLAGSHIP_HARD_2.png'})
+FLEET_DETAIL_ENTER_HARD_1 = Button(area={'cn': (705, 173, 791, 259), 'en': (705, 173, 791, 259), 'jp': (705, 173, 791, 259), 'tw': (705, 173, 791, 259)}, color={'cn': (63, 64, 62), 'en': (63, 64, 62), 'jp': (63, 64, 62), 'tw': (63, 64, 62)}, button={'cn': (705, 173, 791, 259), 'en': (705, 173, 791, 259), 'jp': (705, 173, 791, 259), 'tw': (705, 173, 791, 259)}, file={'cn': './assets/cn/equipment/FLEET_DETAIL_ENTER_HARD_1.png', 'en': './assets/cn/equipment/FLEET_DETAIL_ENTER_HARD_1.png', 'jp': './assets/cn/equipment/FLEET_DETAIL_ENTER_HARD_1.png', 'tw': './assets/cn/equipment/FLEET_DETAIL_ENTER_HARD_1.png'})
+FLEET_DETAIL_ENTER_HARD_2 = Button(area={'cn': (705, 287, 791, 373), 'en': (705, 287, 791, 373), 'jp': (705, 287, 791, 373), 'tw': (705, 287, 791, 373)}, color={'cn': (39, 44, 50), 'en': (39, 44, 50), 'jp': (39, 44, 50), 'tw': (39, 44, 50)}, button={'cn': (705, 287, 791, 373), 'en': (705, 287, 791, 373), 'jp': (705, 287, 791, 373), 'tw': (705, 287, 791, 373)}, file={'cn': './assets/cn/equipment/FLEET_DETAIL_ENTER_HARD_2.png', 'en': './assets/cn/equipment/FLEET_DETAIL_ENTER_HARD_2.png', 'jp': './assets/cn/equipment/FLEET_DETAIL_ENTER_HARD_2.png', 'tw': './assets/cn/equipment/FLEET_DETAIL_ENTER_HARD_2.png'})
 FLEET_ENTER = Button(area={'cn': (502, 474, 517, 489), 'en': (502, 474, 517, 489), 'jp': (502, 474, 517, 489), 'tw': (502, 474, 517, 489)}, color={'cn': (58, 62, 77), 'en': (58, 62, 77), 'jp': (58, 62, 77), 'tw': (58, 62, 77)}, button={'cn': (502, 474, 517, 489), 'en': (502, 474, 517, 489), 'jp': (502, 474, 517, 489), 'tw': (502, 474, 517, 489)}, file={'cn': './assets/cn/equipment/FLEET_ENTER.png', 'en': './assets/en/equipment/FLEET_ENTER.png', 'jp': './assets/jp/equipment/FLEET_ENTER.png', 'tw': './assets/tw/equipment/FLEET_ENTER.png'})
 FLEET_ENTER_FLAGSHIP = Button(area={'cn': (577, 277, 605, 291), 'en': (577, 277, 605, 291), 'jp': (577, 277, 605, 291), 'tw': (577, 277, 605, 291)}, color={'cn': (132, 124, 128), 'en': (132, 124, 128), 'jp': (132, 124, 128), 'tw': (132, 124, 128)}, button={'cn': (577, 277, 605, 291), 'en': (577, 277, 605, 291), 'jp': (577, 277, 605, 291), 'tw': (577, 277, 605, 291)}, file={'cn': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP.png', 'en': './assets/en/equipment/FLEET_ENTER_FLAGSHIP.png', 'jp': './assets/jp/equipment/FLEET_ENTER_FLAGSHIP.png', 'tw': './assets/tw/equipment/FLEET_ENTER_FLAGSHIP.png'})
+FLEET_ENTER_FLAGSHIP_HARD_1 = Button(area={'cn': (593, 173, 681, 261), 'en': (593, 173, 681, 261), 'jp': (593, 173, 681, 261), 'tw': (593, 173, 681, 261)}, color={'cn': (55, 59, 60), 'en': (55, 59, 60), 'jp': (55, 59, 60), 'tw': (55, 59, 60)}, button={'cn': (593, 173, 681, 261), 'en': (593, 173, 681, 261), 'jp': (593, 173, 681, 261), 'tw': (593, 173, 681, 261)}, file={'cn': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP_HARD_1.png', 'en': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP_HARD_1.png', 'jp': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP_HARD_1.png', 'tw': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP_HARD_1.png'})
+FLEET_ENTER_FLAGSHIP_HARD_2 = Button(area={'cn': (594, 287, 681, 374), 'en': (594, 287, 681, 374), 'jp': (594, 287, 681, 374), 'tw': (594, 287, 681, 374)}, color={'cn': (39, 45, 47), 'en': (39, 45, 47), 'jp': (39, 45, 47), 'tw': (39, 45, 47)}, button={'cn': (594, 287, 681, 374), 'en': (594, 287, 681, 374), 'jp': (594, 287, 681, 374), 'tw': (594, 287, 681, 374)}, file={'cn': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP_HARD_2.png', 'en': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP_HARD_2.png', 'jp': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP_HARD_2.png', 'tw': './assets/cn/equipment/FLEET_ENTER_FLAGSHIP_HARD_2.png'})
+FLEET_ENTER_HARD_1 = Button(area={'cn': (909, 173, 996, 260), 'en': (909, 173, 996, 260), 'jp': (909, 173, 996, 260), 'tw': (909, 173, 996, 260)}, color={'cn': (43, 47, 52), 'en': (43, 47, 52), 'jp': (43, 47, 52), 'tw': (43, 47, 52)}, button={'cn': (909, 173, 996, 260), 'en': (909, 173, 996, 260), 'jp': (909, 173, 996, 260), 'tw': (909, 173, 996, 260)}, file={'cn': './assets/cn/equipment/FLEET_ENTER_HARD_1.png', 'en': './assets/cn/equipment/FLEET_ENTER_HARD_1.png', 'jp': './assets/cn/equipment/FLEET_ENTER_HARD_1.png', 'tw': './assets/cn/equipment/FLEET_ENTER_HARD_1.png'})
+FLEET_ENTER_HARD_2 = Button(area={'cn': (909, 287, 996, 374), 'en': (909, 287, 996, 374), 'jp': (909, 287, 996, 374), 'tw': (909, 287, 996, 374)}, color={'cn': (38, 44, 49), 'en': (38, 44, 49), 'jp': (38, 44, 49), 'tw': (38, 44, 49)}, button={'cn': (909, 287, 996, 374), 'en': (909, 287, 996, 374), 'jp': (909, 287, 996, 374), 'tw': (909, 287, 996, 374)}, file={'cn': './assets/cn/equipment/FLEET_ENTER_HARD_2.png', 'en': './assets/cn/equipment/FLEET_ENTER_HARD_2.png', 'jp': './assets/cn/equipment/FLEET_ENTER_HARD_2.png', 'tw': './assets/cn/equipment/FLEET_ENTER_HARD_2.png'})
 FLEET_NEXT = Button(area={'cn': (1234, 327, 1254, 356), 'en': (1234, 327, 1254, 356), 'jp': (1234, 327, 1254, 356), 'tw': (1234, 327, 1254, 356)}, color={'cn': (72, 93, 125), 'en': (72, 93, 125), 'jp': (72, 93, 125), 'tw': (72, 93, 125)}, button={'cn': (1234, 327, 1254, 356), 'en': (1234, 327, 1254, 356), 'jp': (1234, 327, 1254, 356), 'tw': (1234, 327, 1254, 356)}, file={'cn': './assets/cn/equipment/FLEET_NEXT.png', 'en': './assets/en/equipment/FLEET_NEXT.png', 'jp': './assets/jp/equipment/FLEET_NEXT.png', 'tw': './assets/tw/equipment/FLEET_NEXT.png'})
 FLEET_PREV = Button(area={'cn': (26, 327, 46, 356), 'en': (26, 327, 46, 356), 'jp': (26, 327, 46, 356), 'tw': (26, 327, 46, 356)}, color={'cn': (72, 93, 125), 'en': (72, 93, 125), 'jp': (72, 93, 125), 'tw': (72, 93, 125)}, button={'cn': (26, 327, 46, 356), 'en': (26, 327, 46, 356), 'jp': (26, 327, 46, 356), 'tw': (26, 327, 46, 356)}, file={'cn': './assets/cn/equipment/FLEET_PREV.png', 'en': './assets/en/equipment/FLEET_PREV.png', 'jp': './assets/jp/equipment/FLEET_PREV.png', 'tw': './assets/tw/equipment/FLEET_PREV.png'})
 OCR_FLEET_INDEX = Button(area={'cn': (957, 113, 987, 163), 'en': (957, 113, 987, 163), 'jp': (957, 113, 987, 163), 'tw': (957, 113, 987, 163)}, color={'cn': (36, 54, 86), 'en': (36, 54, 86), 'jp': (36, 54, 86), 'tw': (36, 54, 86)}, button={'cn': (957, 113, 987, 163), 'en': (957, 113, 987, 163), 'jp': (957, 113, 987, 163), 'tw': (957, 113, 987, 163)}, file={'cn': './assets/cn/equipment/OCR_FLEET_INDEX.png', 'en': './assets/en/equipment/OCR_FLEET_INDEX.png', 'jp': './assets/jp/equipment/OCR_FLEET_INDEX.png', 'tw': './assets/tw/equipment/OCR_FLEET_INDEX.png'})
